[fix] change float to double

+++ includes/hit.h
@@ -7,9 +7,9 @@
 
 int	update_hit(t_ray *ray, t_ray new);
 
-t_ray	make_hit_sphere(t_sphere *sphere, float *t, t_ray *ray);
-t_ray	make_hit_plane(t_plane *plane, float t, t_ray *ray);
-t_ray	make_hit_cylinder(t_cylinder *cylinder, float t, t_ray *ray);
+t_ray	make_hit_sphere(t_sphere *sphere, double *t, t_ray *ray);
+t_ray	make_hit_plane(t_plane *plane, double t, t_ray *ray);
+t_ray	make_hit_cylinder(t_cylinder *cylinder, double t, t_ray *ray);
