+++ src/discriminant/hit.c
@@ -17,7 +17,7 @@ int	update_hit(t_ray *ray, t_ray new)
 }
 
 //sphere
-t_ray	make_hit_sphere(t_sphere *sphere, float *t, t_ray *ray)
+t_ray	make_hit_sphere(t_sphere *sphere, double *t, t_ray *ray)
 {
 	t_ray	new;
 
@@ -42,7 +42,7 @@ t_ray	make_hit_sphere(t_sphere *sphere, float *t, t_ray *ray)
 	return (new);
 }
 
-t_ray	make_hit_plane(t_plane *plane, float t, t_ray *ray)
+t_ray	make_hit_plane(t_plane *plane, double t, t_ray *ray)
 {
 	t_ray	new;
 
@@ -63,7 +63,7 @@ t_ray	make_hit_plane(t_plane *plane, float t, t_ray *ray)
 	return (new);
 }
 
-t_ray	make_hit_cylinder(t_cylinder *cylinder, float t, t_ray *ray)
+t_ray	make_hit_cylinder(t_cylinder *cylinder, double t, t_ray *ray)
 {
 	t_ray	new;
 	t_vec	r_n_vector;
