+++ src/ray/trace.c
@@ -6,7 +6,7 @@
 
 #include <stdio.h>
 
-t_ray	init_ray(float x, float y, float z)
+t_ray	init_ray(double x, double y, double z)
 {
 	t_ray	ray;
 	t_vec	vec;
@@ -56,14 +56,14 @@ int	is_ray_hit_object(t_object_list *obs, t_ray *ray)
 	return (ray->hit.ray_hit);
 }
 
-int	is_object_visible(t_scene *s, float u, float v, t_ray *ray)
+int	is_object_visible(t_scene *s, double u, double v, t_ray *ray)
 {
 	// t_ray	ray;
-	float	hori_r;
-	float	vert_r;
-	vert_r = (0.5 / (float)s->camera.win.width) + u;
+	double	hori_r;
+	double	vert_r;
+	vert_r = (0.5 / (double)s->camera.win.width) + u;
 	vert_r *= s->camera.viewport_w;
-	hori_r = (0.5 / (float)s->camera.win.height) + v;
+	hori_r = (0.5 / (double)s->camera.win.height) + v;
 	hori_r *= s->camera.viewport_h;
 
 	//ray.vec = vunit(vplus(vplus(s->camera.left_bottom, vmulti_f(s->camera.ver, vert_r)), vmulti_f(s->camera.hor, hori_r)));
@@ -77,7 +77,7 @@ int	is_light(t_light light, t_ray ray)
 	t_vec	light_vec;
 	t_vec	obj_nvec;
 
-	light_vec = vunit(vminus(ray.hit.hit_point, light.coordinate));
+	light_vec = vunit(vminus(ray.hit.hit_point, light.coordinate)); // LP
 	if (ray.hit.in_object) // 빛이 물체 안에 있는지 밖에 있는지 판별 필요 
 		obj_nvec = vmulti_f(ray.hit.hit_normal, -1);
 	else
@@ -94,8 +94,8 @@ int	is_light(t_light light, t_ray ray)
 
 void	shoot_ray(t_mlx *m, t_scene *scene)
 {
-	float	u;
-	float	v;
+	double	u;
+	double	v;
 	int		i;
 	int		j;
 
@@ -107,8 +107,8 @@ void	shoot_ray(t_mlx *m, t_scene *scene)
 		i = 0;
 		while (i < scene->camera.win.width - 1)
 		{
-			u = (float)i / (float)(scene->camera.win.width);
-			v = (float)j / (float)(scene->camera.win.height);
+			u = (double)i / (double)(scene->camera.win.width);
+			v = (double)j / (double)(scene->camera.win.height);
 			if (is_object_visible(scene, u, v, &ray))
 			{
 				if (is_light(scene->light, ray))
@@ -130,8 +130,8 @@ void	shoot_ray(t_mlx *m, t_scene *scene)
 
 // void	shoot_ray(t_mlx *m, t_scene *scene)
 // {
-// 	float	u;
-// 	float	v;
+// 	double	u;
+// 	double	v;
 // 	int		i;
 // 	int		j;
 
@@ -143,8 +143,8 @@ void	shoot_ray(t_mlx *m, t_scene *scene)
 // 		i = 0;
 // 		while (i < scene->camera.win.width - 1)
 // 		{
-// 			u = (float)i / (float)(scene->camera.win.width);
-// 			v = (float)j / (float)(scene->camera.win.height);
+// 			u = (double)i / (double)(scene->camera.win.width);
+// 			v = (double)j / (double)(scene->camera.win.height);
 // 			if (is_object_visible(scene, u, v, &ray))
 // 				my_mlx_pixel_put(&(m->data), i, scene->camera.win.height - 1 - j, to_rgb(vmulti_f(vmin(vec3(1, 1, 1), vplus(get_pixel_ambient_color(scene, ray.hit.hit_color), get_pixel_diffuse_color(scene, &ray))), 255)));
 // 			i++;
